fix: restricted rc_configuration model to be non versioned with no version
